requirements.txt
pytest
selene 2.0.0b17
allure-pytest
